requirements.txt
Flask==3.0.3
requests==2.32.3
GitPython==3.1.43
